Made the download button its own plugin
