+++ plugins/file-highlight/index.html
@@ -34,13 +34,6 @@
 	If, however, the language cannot be determined from the file extension or the file extension is incorrect, you may specify a language as well (with the usual class name way).</p>
 
 	<p>Please note that the files are fetched with XMLHttpRequest. This means that if the file is on a different origin, fetching it will fail, unless CORS is enabled on that website.</p>
-
-	<p>
-		When used in conjunction with the <a href="plugins/toolbar">Toolbar plugin</a>, this plugin can also display a button to download the file.
-		To use it, add a <code>data-download-link</code> attribute on the <code>&lt;pre></code> element.<br />
-		Optionally, the text can also be customized by using a <code>data-download-link-label</code> attribute.
-	</p>
-	<pre><code>&lt;pre data-src="myfile.js" data-download-link data-download-link-label="Download this file">&lt;/pre></code></pre>
 </section>
 
 <section>
@@ -55,9 +48,6 @@
 	<p>File that doesn’t exist:</p>
 	<pre data-src="foobar.js"></pre>
 
-	<p>With a download button:</p>
-	<pre data-src="plugins/file-highlight/prism-file-highlight.js" data-download-link data-download-link-label="Download this file"></pre>
-
 	<p>For more examples, browse around the Prism website. Most large code samples are actually files fetched with this plugin.</p>
 </section>
 
@@ -65,7 +55,6 @@
 
 <script src="prism.js"></script>
 <!-- The File Highlight plugin is already included into Prism.js build -->
-<script src="plugins/toolbar/prism-toolbar.js"></script>
 <script src="scripts/utopia.js"></script>
 <script src="components.js"></script>
 <script src="scripts/code.js"></script>
